requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==1.7.0
certifi==2022.5.18.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gcloud==0.17.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
iniconfig==1.1.1
jws==0.1.3
oauth2client==3.0.0
packaging==21.3
pluggy==1.0.0
protobuf==4.21.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycryptodome==3.4.3
PyJWT==2.4.0
pyparsing==3.0.9
Pyrebase==3.0.27
pytest==7.1.3
pytest-django==4.5.2
python-jwt==2.0.1
pytz==2022.2.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
whitenoise==6.2.0
